sq - split controllers
