+++ src/test/java/hello/HomePageTest.java
@@ -17,11 +17,14 @@ import org.springframework.http.MediaType;
 import org.springframework.test.context.junit4.SpringRunner;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.request.MockMvcRequestBuilders;
+
+import hello.controllers.HomeController;
+
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.xpath;
 
 
 @RunWith(SpringRunner.class)
-@WebMvcTest(WebController.class)
+@WebMvcTest(HomeController.class)
 public class HomePageTest {
